+++ src/integrationTest/java/uk/gov/hmcts/reform/blobrouter/data/SupplierStatementRepositoryTest.java
@@ -6,8 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.test.context.ActiveProfiles;
-import uk.gov.hmcts.reform.blobrouter.data.reconciliation.EnvelopeSupplierStatement;
-import uk.gov.hmcts.reform.blobrouter.data.reconciliation.NewEnvelopeSupplierStatement;
+import uk.gov.hmcts.reform.blobrouter.data.reconciliation.model.EnvelopeSupplierStatement;
+import uk.gov.hmcts.reform.blobrouter.data.reconciliation.model.NewEnvelopeSupplierStatement;
 import uk.gov.hmcts.reform.blobrouter.data.reconciliation.SupplierStatementRepository;
 
 import javax.validation.ClockProvider;
