+++ src/main/java/uk/gov/hmcts/reform/blobrouter/data/reconciliation/model/EnvelopeSupplierStatement.java
@@ -1,24 +1,28 @@
-package uk.gov.hmcts.reform.blobrouter.data.reconciliation;
+package uk.gov.hmcts.reform.blobrouter.data.reconciliation.model;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.UUID;
 
-public class NewEnvelopeSupplierStatement {
+public class EnvelopeSupplierStatement {
 
     public final UUID id;
     public final LocalDate date;
     public final String content;
     public final String contentTypeVersion;
+    public final LocalDateTime createdAt;
 
-    public NewEnvelopeSupplierStatement(
+    public EnvelopeSupplierStatement(
         UUID id,
         LocalDate date,
         String content,
-        String contentTypeVersion
+        String contentTypeVersion,
+        LocalDateTime createdAt
     ) {
         this.id = id;
         this.date = date;
         this.content = content;
         this.contentTypeVersion = contentTypeVersion;
+        this.createdAt = createdAt;
     }
 }
