+++ src/main/java/uk/gov/hmcts/reform/blobrouter/data/reconciliation/model/NewEnvelopeSupplierStatement.java
@@ -1,28 +1,24 @@
-package uk.gov.hmcts.reform.blobrouter.data.reconciliation;
+package uk.gov.hmcts.reform.blobrouter.data.reconciliation.model;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.UUID;
 
-public class EnvelopeSupplierStatement {
+public class NewEnvelopeSupplierStatement {
 
     public final UUID id;
     public final LocalDate date;
     public final String content;
     public final String contentTypeVersion;
-    public final LocalDateTime createdAt;
 
-    public EnvelopeSupplierStatement(
+    public NewEnvelopeSupplierStatement(
         UUID id,
         LocalDate date,
         String content,
-        String contentTypeVersion,
-        LocalDateTime createdAt
+        String contentTypeVersion
     ) {
         this.id = id;
         this.date = date;
         this.content = content;
         this.contentTypeVersion = contentTypeVersion;
-        this.createdAt = createdAt;
     }
 }
